Replace googlemaps api with openstreetmaps api
This commit replaces the use of the googlemaps api with the api for
openstreetmaps.  This removes the need for the user to obtain and supply
an API key for the google api.

The data schema expectations for the return value has also been altered
to accommodate the openstreetmaps response.  See the
[osm wiki](https://wiki.openstreetmap.org/wiki/Nominatim) for details.